pyproject.toml
[project]
name = "ai-based-discovery"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "init-llm>=0.1.3",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langgraph>=0.6.3",
    "mypy>=1.17.0",
    "pygithub>=2.7.0",
    "ruff>=0.12.5",
    "sap-ai-sdk-gen[all]>=5.4.2",
    "sqlalchemy>=2.0.43",
    "databases>=0.9.0",
    "psycopg2-binary>=2.9.10",
    "asyncpg>=0.30.0",
    "structlog>=25.4.0,<26.0.0",
    "langchain-openai>=0.3.29",
    "langchain-anthropic>=0.3.22",
    "requests>=2.32.4",
    "click>=8.2.1",
    "rich>=14.2.0",
]

[project.scripts]
sbs-ai-discovery = "src.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ipython>=9.4.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
